# Pre-allocate memory in `logistic_key`