--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -6,6 +6,9 @@
 ## Features
 - `lorenz_key` and `logistic_key` are more flexible and independent now (one can specify the upper bound of pseudo-random number, hence now they are not limited only to image encryption) ([#39](https://github.com/Saransh-cpp/ChaoticEncryption.jl/pull/39), @Saransh-cpp)
 
+## Optimisations
+- `logistic_key` now pre-allocates the memory, making the function twice as fast as before ([#48](https://github.com/Saransh-cpp/ChaoticEncryption.jl/pull/48), @Saransh-cpp)
+
 ## Documentation
 - Updated all the examples to go with `Julia v1.7.2` migration ([#34](https://github.com/Saransh-cpp/ChaoticEncryption.jl/pull/34), @Saransh-cpp)
 - All the docstrings have been updated to follow `Julia v1.7.2` and to pass the `doctests` ([#32](https://github.com/Saransh-cpp/ChaoticEncryption.jl/pull/32), @Saransh-cpp)